requirements.txt
ray>=2.48.0
PyYAML>=6.0
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
biopython>=1.79
scipy>=1.8.0
networkx>=2.8.0 